Correct namespaces after move
